+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Created documentation for client state
 * Fixed prevention of false-positive handshake timeouts
 * Fixed prevention of false-positive heartbeat timeouts, and fixed the cleanup of heartbeat timeouts
+* Fixed the starting of the heartbeat timeout on client ready
 
 0.8.0
 ===